# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Kaka Portfolio API
  version: "v12.0.0"
  description: API for the Kaka portfolio management system
servers:
  - url: http://114.132.238.12
  - url: http://127.0.0.1:5003
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
...
    AdminMiniBar:
      type: object
      properties:
        date:
          type: string
        value:
          type: integer
        height:
          type: string
        is_latest:
          type: boolean
    AdminRetentionRow:
      type: object
      properties:
        date:
          type: string
        new_users:
          type: integer
        active_users:
          type: integer
        retention_1d:
          type: number
          nullable: true
        retention_3d:
          type: number
          nullable: true
        retention_7d:
          type: number
          nullable: true
        retention_14d:
          type: number
          nullable: true
        retention_30d:
          type: number
          nullable: true
        active_retention_1d:
          type: number
          nullable: true
        active_retention_3d:
          type: number
          nullable: true
        active_retention_7d:
          type: number
          nullable: true
        active_retention_14d:
          type: number
          nullable: true
    AdminAuditItem:
      type: object
      properties:
        id:
          type: integer
        admin_user_id:
          type: string
        admin_username:
          type: string
        action:
          type: string
        target_type:
          type: string
        target_id:
          type: string
        method:
          type: string
        path:
          type: string
        status_code:
          type: integer
        result:
          type: string
        error:
          type: string
        created_at:
          type: string
    AdminUserOpsMetrics:
      type: object
      properties:
        user_total:
          type: integer
        new_today:
          type: integer
        new_7d:
          type: integer
        new_30d:
          type: integer
        dau:
          type: integer
        wau:
          type: integer
        mau:
          type: integer
        last_login_distribution:
          type: object
          properties:
            within_1d:
              type: integer
            within_7d:
              type: integer
            within_30d:
              type: integer
            over_30d:
              type: integer
            never_login:
              type: integer
...
    AdminOverviewResponse:
      type: object
      properties:
        dashboard:
          type: object
          properties:
            new_users_today:
              type: integer
            active_users_today:
              type: integer
            total_users:
              type: integer
            new_users_avatar_count:
              type: integer
            new_user_trend_text:
              type: string
            active_user_trend_text:
              type: string
            new_user_bars:
              type: array
              items:
                $ref: "#/components/schemas/AdminMiniBar"
            active_user_bars:
              type: array
              items:
                $ref: "#/components/schemas/AdminMiniBar"
        retention_rows:
          type: array
          items:
            $ref: "#/components/schemas/AdminRetentionRow"
        users:
          type: object
          properties:
            total:
              type: integer
        user_ops:
          $ref: "#/components/schemas/AdminUserOpsMetrics"
        snapshots:
          type: object
          properties:
            total:
              type: integer
            latest_date:
              type: string
        recent_audits:
          type: array
          items:
            $ref: "#/components/schemas/AdminAuditItem"
    AdminTodoItem:
      type: object
      properties:
        code:
          type: string
        level:
          type: string
        title:
          type: string
        description:
          type: string
        suggestion:
          type: string
    AdminSummaryTodoResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminTodoItem"
        snapshot:
          type: object
          properties:
            active_invites:
              type: integer
            invite_threshold:
              type: integer
            disabled_policies:
              type: integer
            degraded_upstream:
              type: integer
            failed_audits_today:
              type: integer
            disabled_users:
              type: integer
    AdminMetaDictionariesResponse:
      type: object
      properties:
        status_labels:
          type: object
          additionalProperties:
            type: string
        action_labels:
          type: object
          additionalProperties:
            type: string
        policy_labels:
          type: object
          additionalProperties:
            type: string
        policy_impacts:
          type: object
          additionalProperties:
            type: string
        policy_type_labels:
          type: object
          additionalProperties:
            type: string
        register_method_labels:
          type: object
          additionalProperties:
            type: string
        error_labels:
          type: object
          additionalProperties:
            type: string
        config_labels:
          type: object
          additionalProperties:
            type: string
...
paths:
...
  /api/admin/meta/dictionaries:
    get:
      operationId: apiAdminMetaDictionariesGet
      summary: 后台字典与标签
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminMetaDictionariesResponse"
        "401":
          description: 未登录或登录过期
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: 无后台权限或账号不可用
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/admin/overview:
    get:
      operationId: apiAdminOverviewGet
      summary: 后台概览看板
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: force
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOverviewResponse"
        "400":
          description: force 参数不合法
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: 未登录或登录过期
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: 无后台权限或账号不可用
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/admin/summary/todo:
    get:
      operationId: apiAdminSummaryTodoGet
      summary: 后台待办与巡检摘要
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: invite_threshold
          required: false
          schema:
            type: integer
            default: 200
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSummaryTodoResponse"
        "400":
          description: invite_threshold 不合法
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: 未登录或登录过期
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: 无后台权限或账号不可用
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...